Adding Offer Part

+++ LunamaroApi/Lunamaroapi/DTOs/Admin/ordersListDTO.cs
@@ -7,7 +7,7 @@ namespace Lunamaroapi.DTOs.Admin
         public int OrderId { get; set; }
         public string CustomerName { get; set; }
         public string PhoneNumber { get; set; }
-        public double totalAmount { get; set; }
+        public decimal totalAmount { get; set; }
         public OrderStatus orderStatus { get; set; }
         public DateTime OrderDate { get; set; }
         public PaymentType paymentType { get; set; }
