+++ LunamaroApi/Lunamaroapi/DTOs/DashBoard/HeaderStatsDTO.cs
@@ -4,7 +4,7 @@
     {
         public int TotalMenuItems { get; set; }
         public int TotalUsers { get; set; }
-        public Double TotalRevenue { get; set; }
+        public decimal TotalRevenue { get; set; }
         public int TotalOrders { get; set; }
     }
 }
