+++ LunamaroApi/Lunamaroapi/DTOs/Order/OrderItemDTO.cs
@@ -7,9 +7,9 @@
         public string ItemName { get; set; }
         public string? ImageUrl { get; set; }
         public string Description { get; set; }
-        public double UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public int Quantity { get; set; }
 
-        public double Total => UnitPrice * Quantity;
+        public decimal Total => UnitPrice * Quantity;
     }
 }
